modified script > i3-exit.sh
changed: replace the cumbersome dbus-send commands with systemctl calls. Since
we use systemd either way and since it can handle privileges & shutdown
options, we might as well just use it.

added script > i3-display.sh

This script simply queries xrandr and if it finds an external (analogue)
monitor connected, it will set both the internal and external monitor to their
preferred settings (--auto). And no matter whether another monitor is
connected or not, this script will always set the internal monitor as primary
(which is important for the i3wm statusbar to work appear on that
monitor). This script is the manual option in case this needs to be set up
again, the normal startup for i3wm will read the exact same content from
~/.xinitrc

added script > i3-wallpaper.sh

This script uses the same probing syntax as the script above to figure out
whether we have one or two monitors connected. It then uses nitrogen to set
wallpapers accordingly (with a possibly more public-friendly single-output
wallpaper).