added script > eve-intel.sh
This EVE-related script is used to scrape zkillboard.org killmails involving
supers and titans for a number of different timeframes in order to monitor the
general use of these shiptypes. It checks for stale data, uses temporary files
(located in /dev/shm/) and downloads them as necessary. This script requires
json from https://github.com/trentm/json to work properly, since it is used to
parse the zkill killmail syntax.