added script > eve-watchlist_wrapper.sh
This script works as a wrapper around eve-watchlist.rkt, generating a number
of different watchlist templates with different contents and different lengths
automatically. This script is meant to be run via cronjob and therefore
includes a "latest" command (run by default) that can be run at arbitrary
intervals, as well as a "delayed" command (-d on the command-line) that can be
used to choose a longer period between executions while still using the same script.